+++ Domain/Entities/Services/Service.cs
@@ -8,7 +8,7 @@ public class Service : Entity
     public required string Description { get; set; }
     public required ICollection<int> RequiredHealthCertificateTemplateIds { get; set; }
     public required ICollection<int> ResultHealthCertificateTemplateIds { get; set; }
-    public required float PricePerHourForMaterials { get; set; }
-    public required float PricePerHourForEquipment { get; set; }
-    public required TimeSpan Duration { get; set; }
+    public required decimal PricePerHourForMaterials { get; set; }
+    public required decimal PricePerHourForEquipment { get; set; }
+    public required TimeSpan Duration { get; set; } 
 }
